+++ src/bika/lims/interfaces/__init__.py
@@ -829,17 +829,6 @@ class IGetDefaultFieldValueARAddHook(Interface):
     """
 
 
-class IGetStickerTemplates(Interface):
-    """Marker interface to get stickers for a specific content type.
-
-    An IGetStickerTemplates adapter should return a result with the
-    following format:
-
-    :return: [{'id': <template_id>,
-             'title': <template_title>}, ...]
-    """
-
-
 class IARReport(Interface):
     """Marker interface for published AR Reports
     """
