+++ src/senaite/core/interfaces/__init__.py
@@ -22,6 +22,7 @@ from plone.app.z3cform.interfaces import IPloneFormLayer
 from plone.protect.interfaces import IDisableCSRFProtection
 from senaite.core.interfaces.catalog import *  # noqa:F401,F403
 from senaite.core.interfaces.datamanager import IDataManager  # noqa:F401
+from senaite.core.interfaces.stickers import *  # noqa:F401,F403
 from senaite.core.interfaces.widget import *  # noqa:F401,F403
 from zope.interface import Interface
